Initial refactor of tests.
The rework of tests includes adding a new Rv_Session class, whose
job is to provide abstraction for a connection between client and
guest and give some convenience methods for frequently used tasks.

It also includes update to spice_utils.py, which is now meant as
a library for functions, which are not necessarily directly
involved with connection, but can be used to manipulate different
virtual machines.

rv_connect.py and rv_fullscreen are the first tests to use the
updated class/utils. More tests will follow.

Note that config files do not yet reflect the changes made

Signed-off-by: Tomas Jamrisko <tjamrisk@redhat.com>